Split file ReadBoundaryDataH5
to files:
- WorldtubeDataManager
- WorldtubeBufferUpdater
