+++ tests/Unit/Evolution/Systems/Cce/CMakeLists.txt
@@ -21,11 +21,11 @@ set(LIBRARY_SOURCES
   Test_OptionTags.cpp
   Test_PreSwshDerivatives.cpp
   Test_PrecomputeCceDependencies.cpp
-  Test_ReadBoundaryDataH5.cpp
   Test_ScriPlusInterpolationManager.cpp
   Test_ScriPlusValues.cpp
   Test_SwshDerivatives.cpp
   Test_Tags.cpp
+  Test_WorldtubeData.cpp
   )
 
 add_test_library(
