+++ tests/Unit/Evolution/Systems/Cce/Test_Tags.cpp
@@ -7,8 +7,8 @@
 #include <type_traits>
 
 #include "DataStructures/DataBox/DataBox.hpp"
-#include "Evolution/Systems/Cce/ReadBoundaryDataH5.hpp"
 #include "Evolution/Systems/Cce/Tags.hpp"
+#include "Evolution/Systems/Cce/WorldtubeBufferUpdater.hpp"
 #include "Helpers/DataStructures/DataBox/TestHelpers.hpp"
 
 /// \cond
